tests: extract diagnostics core from fx3_cmd into fx3_core (Phase 1)
Behavior-preserving split of the fx3_cmd monolith. Move the operator
-diagnostics primitives into a new shared translation unit so both the
firmware test harness (this repo) and the rx888-tools diagnostics CLI
can link the same core.

New tests/fx3_core.{c,h}: firmware upload (find_rx888_stream,
upload_firmware), the vendor-command primitives (do_test/gpio/adc/att/
vga/wdg_max/start/stop/i2cr/i2cw/reset/usbreset/reload/raw), and the
read-only status reads (do_test_stack_check/stats_pll/stats_shdn).
Functions gain external linkage and prototypes in fx3_core.h.

tests/fx3_cmd.c: include fx3_core.h, drop the moved bodies and their
redundant forward declarations; keep the regression/fuzz/soak harness,
the interactive debug console, and main()/dispatch. do_stats was
already shared in fx3_stats.c.

tests/Makefile: add fx3_core.o/fx3_core.h to the build.

Verified: clean -Wall -Wextra build; fx3_cmd usage and dispatch tokens
identical to pre-refactor; make check-health passes.
